backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
deepgram-sdk
pdfplumber
python-docx